// src/providerRuntime/types.ts
export type Error =
    | { type: "unsupported_request" }
    | {
        type: "validation_error";

        /**
         * List of errors, so all fields that failed validation can
         * be highlighted at once.
         */
        errors: Array<ValidationError>;
    }
    | { type: "http"; error: HttpRequestError }
    | { type: "data"; message: string }
    | { type: "deserialization"; message: string }
    | { type: "config"; message: string }
    | { type: "not_found" }
    | { type: "proxy_disconnected" }
    | { type: "invocation"; message: string }
    | { type: "other"; message: string };
// ...
export type HttpRequestError =
    | { type: "offline" }
    | { type: "no_route" }
    | { type: "connection_refused" }
    | { type: "timeout" }
    | { type: "response_too_big" }
    | { type: "server_error"; statusCode: number; response: Uint8Array }
    | { type: "other"; reason: string };
// ...
export type ValidationError = {
    /**
     * Refers to a field from the query schema.
     */
    fieldName: string;

    /**
     * Description of why the validation failed.
     */
    message: string;
};